+++ CHANGELOG.md
@@ -8,9 +8,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Added
 
+- Support for [netrc] authentication.
 - Prebuilt linux-riscv64 binary release.
 - Support for CRSH greeting message format 2.
 - Sending of config errors/warnings to ccache.
+
+[netrc]: https://everything.curl.dev/usingcurl/netrc.html
 
 ### Fixed